chore: set up installable helper library
