+++ apps/showcase-e2e/src/support/index.ts
@@ -15,3 +15,4 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands';
+import './namespace';
