+++ libs/cypress-indexeddb/README.md
@@ -1,0 +1,7 @@
+# cypress-indexeddb
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test cypress-indexeddb` to execute the unit tests via [Jest](https://jestjs.io).
