+++ libs/cypress-indexeddb/jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  displayName: 'cypress-indexeddb',
+  preset: '../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/libs/cypress-indexeddb',
+};
