+++ libs/cypress-indexeddb/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@this-dot/cypress-indexeddb",
+  "version": "0.0.1"
+}
