+++ libs/cypress-indexeddb/src/lib/cypress-indexeddb.spec.ts
@@ -1,0 +1,7 @@
+// import { cypressIndexeddb } from './cypress-indexeddb';
+//
+// describe('cypressIndexeddb', () => {
+//   it('should work', () => {
+//     expect(cypressIndexeddb()).toEqual('cypress-indexeddb');
+//   });
+// });
